feat: Реализация RBAC, переработка UserController и сервисов, подготовка к логированию и обработке ошибок
- Созданы 3 дополнительных контроллера и соответствующие интерфейсы + реализации для поддержки RBAC.
- Введена иерархия ролей: USER, ADMIN_ONE, ADMIN_TWO, ADMIN_THREE.
- Переработано создание и обновление пользователей с новым SQL, соответствующим RBAC.
- UserController теперь ограничен функционалом для роли USER:
- Можно читать, обновлять и удалять только свои данные.
- Добавлен универсальный ResponseBuilder с использованием фабричного паттерна для формирования ответов API.
- Значительно переработаны DTO:
- Упрощены для пользователя.
- Добавлены расширенные DTO для админских ролей с дополнительными данными.
- UserService и CRUD сервисы переписаны:
- Убраны лишние методы.
- Логика максимально оптимизирована.
- Временно удалены mapStruct мапперы — переход на ручной маппинг в CRUD сервисах.
- Планируется вернуть mapStruct позже для удобства и производительности.

+++ src/main/resources/sqlQuery/delete_user_by_id.sql
@@ -1,1 +1,17 @@
-delete from users where id = ?;
+with
+delete_user_role_users as(
+delete from user_role where user_id = ?
+returning user_id
+),
+delete_profile as(
+delete from profiles
+where user_id = (select user_id from delete_user_role_users)
+),
+delete_emails as(
+delete from emails
+where user_id = (select user_id from delete_user_role_users)
+),
+delete_users as(
+delete from users where id = (select user_id from delete_user_role_users)
+)
+select user_id from delete_user_role_users;
